hide menu from dom when not showing
previously the menu was hidden rotating it 90 degrees from
the X axis, with the intention to animate it. I decided I won't
animate the menu toggle, but left the rotation in the properties,
which may be accessed by keyboard by accident. This time the menu
is hidden with the property display:none, to prevent any kind of
unforeseen behaviors.