change checkbox id to always work with labels
the previous method used the to-do title as ID so the labels
know which one it belongs, but repeated names caused the labels
to interact with the wrong checkbox. Now each ID will use all
the necessary indexes to fix this issue, and names can be
repeated without causing further problems.