move separation margins to class, optimization
Previously the separation when dragging a card over it was a
hardcoded value, but now it's on a new class. This way this
property can be removed only for the card with the separation
value instead for every single card on the list.
This change reduced the CPU usage from max 8% to max 3% (only
happening if the user drags very fast over all the cards).